backport of commit e4e70b086a304e4e80496d95f514a4de6e926b3b

+++ drivers/shared/executor/procstats/list_default.go
@@ -32,7 +32,7 @@ func List(executorPID int) *set.Set[ProcessID] {
 		nextPPID := stack.Pop()
 		result.Insert(ProcessID(nextPPID))
 
-		p, err := process.NewProcessWithContext(ctx, int32(nextPPID))
+		p, err := process.NewProcessWithContext(ctx, nextPPID)
 		if err != nil {
 			continue
 		}
